Test signature proof generation by a notary.
